mark day action, which will be used when performing a day action (mark/clear)

+++ src/app/(frontend)/home/components/calendar/MonthDayGrid/markDay.action.ts
@@ -1,6 +1,6 @@
 "use server";
 import { revalidatePath } from "next/cache";
-import { DBDayMark } from "../../../../../types/dateTypes";
+import { DBDayMark } from "@/app/types/dateTypes";
 import { serverSideSupabase } from "@/app/(frontend)/supabaseServerShit/serverSideSupabase";
 
 export type ActionSettings = {
@@ -19,38 +19,24 @@ export const markDayAction = async (
 
     const user_id = (await supabase.auth.getUser()).data.user?.id;
 
-    if (typeof date === "string") {
+    if (typeof date === "string" && user_id) {
         // delete single
         if (actionSettings.clear && !multipleDatesArr) {
-            console.log("deleting single");
-            const { data, error } = await supabase.from("marked_day").delete().eq("date", date);
+            await supabase.from("marked_day").delete().eq("user_id", user_id).eq("date", date);
         }
         // delete multiple
         if (actionSettings.clear && multipleDatesArr) {
-            console.log("deleting multiple");
-            const insertData = multipleDatesArr.map((date) => ({
-                date,
-                day_mark_id: selectedDayMark?.id,
-                user_id,
-            }));
-            const { data, error } = await supabase.from("marked_day").delete().in("date", multipleDatesArr);
-            console.log(error);
+            await supabase.from("marked_day").delete().eq("user_id", user_id).in("date", multipleDatesArr);
         }
         // create single
         if (!actionSettings.clear && !multipleDatesArr) {
-            console.log("creating single");
-
             if (actionSettings.overwrite) {
-                console.log("overwrite triggered");
-                const { data, error } = await supabase.from("marked_day").delete().eq("date", date);
+                await supabase.from("marked_day").delete().eq("user_id", user_id).eq("date", date);
             }
-            const { data, error } = await supabase
-                .from("marked_day")
-                .insert([{ date: date, day_mark_id: selectedDayMark?.id, user_id }]);
+            await supabase.from("marked_day").insert([{ date: date, day_mark_id: selectedDayMark?.id, user_id }]);
         }
         // create multiple
         if (!actionSettings.clear && multipleDatesArr) {
-            console.log("creating multiple");
             const insertData = multipleDatesArr.map((date) => ({
                 date,
                 day_mark_id: selectedDayMark?.id,
@@ -58,16 +44,21 @@ export const markDayAction = async (
             }));
 
             if (actionSettings.overwrite) {
-                console.log("overwrite triggered");
-                await supabase.from("marked_day").delete().in("date", multipleDatesArr);
+                // overwrite - delete and insert new
+                await supabase.from("marked_day").delete().eq("user_id", user_id).in("date", multipleDatesArr);
                 await supabase.from("marked_day").insert(insertData);
             } else {
-                const { data, error } = await supabase.from("marked_day").select("date").in("date", multipleDatesArr);
-                console.log(data);
+                // non overwrite - filterout any pre-existing dates before inserting
+                const { data } = await supabase
+                    .from("marked_day")
+                    .select("date")
+                    .eq("user_id", user_id)
+                    .in("date", multipleDatesArr);
+
                 const filteredData = insertData.filter(
                     (markedDay) => !data?.find((marked_day: { date: string }) => marked_day.date === markedDay.date)
                 );
-                console.log(filteredData);
+
                 await supabase.from("marked_day").insert(filteredData);
             }
         }
